Update README.md
We have built a line follower robot that is able to follow both Black and White Lines. Also, the LFR is able to take Right and Left Turn on both the Black and White lines. 
LFR has been able to shift the path from Black to White or vice versa. 

REQUIRED HARDWARE :
1. Arduino UNO
2. L298 Motor Drive
3. 2 * N20 Gear Motor
4. 2 * Rubber Wheel
5. 12-volt Battery
6.  jumper wire
7. 5 - Channel IR sensor

CHASSIS: Design the chassis in Creo-parametric with the help of component dimensions. Cut the chassis with the help of a CNC laser cutter. The material used for building the chassis is a Transparent Acrylic Sheet.

KEY SKILL : Arduino Programming, Cero Parametric Modeling